# docs/system/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AuditLogSearchResponse": {
        "description": "Envelope for audit log search results.",
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/AuditLogRead"
            },
            "title": "Results",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "results",
          "total"
        ],
        "title": "AuditLogSearchResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "HolidayCreate": {
        "description": "Payload for creating holidays.",
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "date",
          "name"
        ],
        "title": "HolidayCreate",
        "type": "object"
      },
...
      "SystemConfigurationRead": {
        "description": "Full configuration view returned to clients.",
        "properties": {
          "approval_escalation_hours": {
            "default": 24,
            "minimum": 0.0,
            "title": "Approval Escalation Hours",
            "type": "integer"
          },
          "auto_cancel_pending_hours": {
            "default": 48,
            "minimum": 0.0,
            "title": "Auto Cancel Pending Hours",
            "type": "integer"
          },
          "booking_lead_time_hours": {
            "default": 4,
            "minimum": 0.0,
            "title": "Booking Lead Time Hours",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "holidays": {
            "items": {
              "$ref": "#/components/schemas/HolidayRead"
            },
            "title": "Holidays",
            "type": "array"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "maintenance_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maintenance Message"
          },
          "maintenance_mode": {
            "default": false,
            "title": "Maintenance Mode",
            "type": "boolean"
          },
          "max_active_bookings_per_user": {
            "default": 2,
            "minimum": 0.0,
            "title": "Max Active Bookings Per User",
            "type": "integer"
          },
          "max_pending_bookings_per_user": {
            "default": 3,
            "minimum": 0.0,
            "title": "Max Pending Bookings Per User",
            "type": "integer"
          },
          "require_booking_approval": {
            "default": true,
            "title": "Require Booking Approval",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "working_day_end": {
            "default": "18:00:00",
            "format": "time",
            "title": "Working Day End",
            "type": "string"
          },
          "working_day_start": {
            "default": "08:00:00",
            "format": "time",
            "title": "Working Day Start",
            "type": "string"
          },
          "working_days": {
            "items": {
              "type": "string"
            },
            "title": "Working Days",
            "type": "array"
          },
          "working_hours": {
            "items": {
              "$ref": "#/components/schemas/WorkingHourRead"
            },
            "title": "Working Hours",
            "type": "array"
          }
        },
        "required": [
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "SystemConfigurationRead",
        "type": "object"
      },
      "SystemConfigurationUpdate": {
        "description": "Patch payload for system configuration updates.",
        "properties": {
          "approval_escalation_hours": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Escalation Hours"
          },
          "auto_cancel_pending_hours": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Cancel Pending Hours"
          },
          "booking_lead_time_hours": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Booking Lead Time Hours"
          },
          "maintenance_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maintenance Message"
          },
          "maintenance_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Maintenance Mode"
          },
          "max_active_bookings_per_user": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Active Bookings Per User"
          },
          "max_pending_bookings_per_user": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pending Bookings Per User"
          },
          "require_booking_approval": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Require Booking Approval"
          },
          "working_day_end": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Day End"
          },
          "working_day_start": {
            "anyOf": [
              {
                "format": "time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Day Start"
          },
          "working_days": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Days"
          }
        },
        "title": "SystemConfigurationUpdate",
        "type": "object"
      },
...
      "UserActivityReport": {
        "description": "Activity report response payload.",
        "properties": {
          "entries": {
            "items": {
              "$ref": "#/components/schemas/UserActivityEntry"
            },
            "title": "Entries",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          }
        },
        "required": [
          "entries"
        ],
        "title": "UserActivityReport",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "\u0e23\u0e30\u0e1a\u0e1a\u0e08\u0e2d\u0e07\u0e23\u0e16\u0e2a\u0e33\u0e19\u0e31\u0e01\u0e07\u0e32\u0e19 - Vehicle booking and management system API",
    "title": "Office Vehicle Booking System API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/system/activity": {
      "get": {
        "description": "Return user activity metrics derived from audit logs.",
        "operationId": "user_activity_report_endpoint_api_v1_system_activity_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 7,
              "maximum": 90,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserActivityReport"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "User Activity Report Endpoint",
        "tags": [
          "system-admin"
        ]
      }
    },
    "/api/v1/system/audit": {
      "get": {
        "description": "Search the audit trail with flexible filters.",
        "operationId": "search_audit_endpoint_api_v1_system_audit_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Id"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Action"
            }
          },
          {
            "in": "query",
            "name": "resource",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Resource"
            }
          },
          {
            "in": "query",
            "name": "status_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Code"
            }
          },
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Query"
            }
          },
          {
            "in": "query",
            "name": "date_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "in": "query",
            "name": "date_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditLogSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Audit Endpoint",
        "tags": [
          "system-admin"
        ]
      }
    },
    "/api/v1/system/audit/statistics": {
      "get": {
        "description": "Return quick statistics for audit trail monitoring.",
        "operationId": "audit_statistics_endpoint_api_v1_system_audit_statistics_get",
        "parameters": [
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 24,
              "minimum": 1,
              "title": "Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "anyOf": [
                      {
                        "type": "integer"
                      },
                      {
                        "additionalProperties": {
                          "type": "integer"
                        },
                        "type": "object"
                      }
                    ]
                  },
                  "title": "Response Audit Statistics Endpoint Api V1 System Audit Statistics Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Audit Statistics Endpoint",
        "tags": [
          "system-admin"
        ]
      }
    },
    "/api/v1/system/config": {
      "get": {
        "description": "Return the current global system configuration.",
        "operationId": "read_system_configuration_api_v1_system_config_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemConfigurationRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read System Configuration",
        "tags": [
          "system-admin"
        ]
      },
      "put": {
        "description": "Update global system settings.",
        "operationId": "update_system_configuration_endpoint_api_v1_system_config_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SystemConfigurationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemConfigurationRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update System Configuration Endpoint",
        "tags": [
          "system-admin"
        ]
      }
    },
...
    "/api/v1/system/holidays": {
      "get": {
        "description": "Return configured organisation-wide holidays.",
        "operationId": "list_holidays_endpoint_api_v1_system_holidays_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HolidayRead"
                  },
                  "title": "Response List Holidays Endpoint Api V1 System Holidays Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Holidays Endpoint",
        "tags": [
          "system-admin"
        ]
      },
      "post": {
        "description": "Create a new holiday entry.",
        "operationId": "create_holiday_endpoint_api_v1_system_holidays_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HolidayCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolidayRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Holiday Endpoint",
        "tags": [
          "system-admin"
        ]
      }
    },
    "/api/v1/system/holidays/{holiday_id}": {
      "delete": {
        "description": "Delete a configured holiday.",
        "operationId": "delete_holiday_endpoint_api_v1_system_holidays__holiday_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "holiday_id",
            "required": true,
            "schema": {
              "title": "Holiday Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Holiday Endpoint",
        "tags": [
          "system-admin"
        ]
      }
    },
    "/api/v1/system/working-hours": {
      "get": {
        "description": "Return configured working hour windows.",
        "operationId": "list_working_hours_endpoint_api_v1_system_working_hours_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/WorkingHourRead"
                  },
                  "title": "Response List Working Hours Endpoint Api V1 System Working Hours Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Working Hours Endpoint",
        "tags": [
          "system-admin"
        ]
      }
    },
    "/api/v1/system/working-hours/{day_of_week}": {
      "delete": {
        "description": "Remove configured working hours for the specified day.",
        "operationId": "delete_working_hours_endpoint_api_v1_system_working_hours__day_of_week__delete",
        "parameters": [
          {
            "in": "path",
            "name": "day_of_week",
            "required": true,
            "schema": {
              "title": "Day Of Week",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Working Hours Endpoint",
        "tags": [
          "system-admin"
        ]
      },
      "put": {
        "description": "Create or update working hours for a given day of week.",
        "operationId": "upsert_working_hours_endpoint_api_v1_system_working_hours__day_of_week__put",
        "parameters": [
          {
            "in": "path",
            "name": "day_of_week",
            "required": true,
            "schema": {
              "title": "Day Of Week",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkingHourCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkingHourRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upsert Working Hours Endpoint",
        "tags": [
          "system-admin"
        ]
      }
    },
...
  }
}